add faucetInteract.js
